Add files via upload
Create a public class named Exponent.  Inside the main method of this class, obtain a double input using a dialog box.
Pass the input value to a non-static method that returns the square value of the parameter.
Display the return value of the method using an output dialog.
Next pass the input value to a non-static method that returns the cube of the parameter.
Display the return value of the method using an output dialog.  Save the class as Exponent.java.