Add files via upload
Create a class named Purchase.  Each Purchase contains an invoice number, amount of sale, and amount of sales tax.
Include set methods for the invoice number and sale amount.  Within the set() method for the sale amount,
calculate the sales tax as 7.5% (using a static field in the Purchase class) of the sale amount.
Also include a display method that displays a purchase's details in a well formatted output display.
Bellevue University